LayoutTests/http/tests/security/cross-origin-css-expected.txt

LINK Cross-origin, HTML, valid: rgb(255, 255, 0)
LINK + IMPORT Cross-origin, HTML, invalid: transparent
LINK Cross-origin, CSS, invalid: rgb(255, 255, 0)
LINK Same-origin, HTML, invalid: rgb(255, 255, 0)
IMPORT Cross-origin, HTML, valid: rgb(255, 255, 0)
IMPORT Cross-origin, CSS, invalid: rgb(255, 255, 0)
IMPORT Same-origin, HTML, invalid: rgb(255, 255, 0)
